% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{formatear_num}
\alias{formatear_num}
\title{Gráfica de formato propio de números}
\usage{
formatear_num(num)
}
\arguments{
\item{num}{es un tipo de dato num\'{e}rico}
}
\value{
el numero a caracter segun su orden
}
\description{
Gráfica de formato propio de números
}
\section{Quasiquotation}{

\code{formatear_num} es una funci\'{o}n \link[rlang:nse-defuse]{quoting function}. Esto significa
que se utiliza en el contexto de un formato de numeros para un label
}

\examples{
formatear_num(10000)
}
\author{
Gerardo Orantes Jordan, \email{gorantes@gerenciadelpoder.com.mx}
}
