% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frecuencia.R
\name{frecuencia_gota}
\alias{frecuencia_gota}
\title{Graficar frecuencia de barras con esquinas suaves estilo GP}
\usage{
frecuencia_gota(bd, x, y, color_base = "#912F40")
}
\arguments{
\item{bd}{el data frame de entrada}

\item{x}{es la variable dependiente, eje y}

\item{y}{es la variable independiente, eje x, a ser la frecuencia}

\item{color_base}{el color del gr\'{a}fico, por defecto es color Gerencia del poder.}
}
\value{
regresa un gráfico de barras ggplot2 a utilizar con frecuencia
}
\description{
Graficar frecuencia de barras con esquinas suaves estilo GP
}
\section{Quasiquotation}{

\code{frecuencia_gota()} es una funci\'{o}n \link[rlang:nse-defuse]{quoting function}. Esto significa
que se utiliza en el contexto del proceso de graficación de elementos ggplot2 y es una gráfica de barras con
esquinas suaves.
}

\author{
Gerardo Orantes Jordan, \email{gorantes@gerenciadelpoder.com.mx}
}
