% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpapply.R
\name{gpapply}
\alias{gpapply}
\title{A Gaussian process prediction reference function applied over a list of features}
\usage{
gpapply(data, x_lst, y_, trainingpect = 0.1, seed = 123, iter = 100,
  chains = 4, conf = 0.5, condition1 = NULL, condition2 = NULL,
  dn = 64, rho_alpha = 4, rho_beta = 4, alpha_mean = 0, alpha_sd = 1,
  sigma_mean = 0, sigma_sd = 1)
}
\arguments{
\item{data}{A list of input values}

\item{x_lst}{a list of names of x variables to be examed, they should match with the data frame variable names}

\item{y_}{name of y variable, should match with the data frame variable name}

\item{trainingpect}{the percent of training instances, the default value is 10\% of the whole sample, avoid high number of training instances when you want the Stan program to be done faster}

\item{seed}{a random generator seed to designate such that your results will be the same across different runs}

\item{iter}{Number of iterations you would like Stan to run}

\item{chains}{Number of chains you would like Stan to utilize}

\item{conf}{Confidence level resulting in the interval displayed for the first derivative of the gaussian process prediction}

\item{condition1}{The first condition you want the first derivative to satisfy, such as ">0", the default value is NULL}

\item{condition2}{The second condition you want the first derivative to satisfy, such as "<2", the default value is NULL}

\item{dn}{The number you want to sample for density calculation for the first derivative's distribution for each x*, the default value is 64}

\item{rho_alpha}{Prior distribution of rho is set to inverse gamma distribution with rho ~ inv_gamma(rho_alpha, rho_beta), default value is inv_gamma(4, 4)}

\item{rho_beta}{Prior distribution of rho is set to inverse gamma distribution with rho ~ inv_gamma(rho_alpha, rho_beta), default value is inv_gamma(4, 4)}

\item{alpha_mean}{Prior distribution of alpha is set to normal distribution with alpha ~ normal(alpha_mean, alpha_sd), default value is normal(0, 1)}

\item{alpha_sd}{Prior distribution of alpha is set to normal distribution with alpha ~ normal(alpha_mean, alpha_sd), default value is normal(0, 1)}

\item{sigma_mean}{Prior distribution of sigma is set to normal distribution with sigma ~ normal(sigma_mean, sigma_sd), default value is normal(0, 1)}

\item{sigma_sd}{Prior distribution of sigma is set to normal distribution with sigma ~ normal(sigma_mean, sigma_sd), default value is normal(0, 1)}
}
\value{
Pdf plots describing Gaussian process prediction and relationship intensity in uncertainty interval and density estimation
}
\description{
A Gaussian process prediction reference function applied over a list of features
}
\examples{
gpapply(data = winequality_red, x_lst = wine_red_x_lst, y_ = 'quality', iter = 1000, condition1 = '>0')
}
