% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpslope.R
\name{gpslope}
\alias{gpslope}
\title{A Gaussian process prediction reference on slopes}
\usage{
gpslope(data, x_, y_, trainingpect = 0.1, seed = 123, iter = 100,
  chains = 4, conf = 0.5, condition1 = NULL, condition2 = NULL,
  dn = 64, rho_alpha = 4, rho_beta = 4, alpha_mean = 0, alpha_sd = 1,
  sigma_mean = 0, sigma_sd = 1)
}
\arguments{
\item{data}{A list of input values}

\item{x_}{name of x variable, should match with the data frame variable name}

\item{y_}{name of y variable, should match with the data frame variable name}

\item{trainingpect}{the percent of training instances, the default value is 10\% of the whole sample, avoid high number of training instances when you want the Stan program to be done faster}

\item{seed}{a random generator seed to designate such that your results will be the same across different runs}

\item{iter}{Number of iterations you would like Stan to run}

\item{chains}{Number of chains you would like Stan to utilize}

\item{conf}{Confidence level resulting in the interval displayed for the first derivative of the gaussian process prediction}

\item{condition1}{The first condition you want the first derivative to satisfy, such as ">0", the default value is NULL}

\item{condition2}{The second condition you want the first derivative to satisfy, such as "<2", the default value is NULL}

\item{dn}{The number you want to sample for density calculation for the first derivative's distribution for each x*, the default value is 64}

\item{rho_alpha}{Prior distribution of rho is set to inverse gamma distribution with rho ~ inv_gamma(rho_alpha, rho_beta), default value is inv_gamma(4, 4)}

\item{rho_beta}{Prior distribution of rho is set to inverse gamma distribution with rho ~ inv_gamma(rho_alpha, rho_beta), default value is inv_gamma(4, 4)}

\item{alpha_mean}{Prior distribution of alpha is set to normal distribution with alpha ~ normal(alpha_mean, alpha_sd), default value is normal(0, 1)}

\item{alpha_sd}{Prior distribution of alpha is set to normal distribution with alpha ~ normal(alpha_mean, alpha_sd), default value is normal(0, 1)}

\item{sigma_mean}{Prior distribution of sigma is set to normal distribution with sigma ~ normal(sigma_mean, sigma_sd), default value is normal(0, 1)}

\item{sigma_sd}{Prior distribution of sigma is set to normal distribution with sigma ~ normal(sigma_mean, sigma_sd), default value is normal(0, 1)}
}
\value{
A list consiting of the Stan fit object, three ggplot plot objects and the user-specified argument values
}
\description{
A Gaussian process prediction reference on slopes
}
\examples{
resultlst <- gpviz(data = hsbReport_cleaned, x_ = 'deposit_asset_ratio', y_ = 'loss_asset_ratio', seed = 123456, iter = 2000, conf = 0.95, condition1 = '>0', condition2 = NULL)
}
