% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpviz-package.R
\docType{package}
\name{gpviz-package}
\alias{gpviz-package}
\alias{gpviz}
\title{The 'gpviz' package.}
\description{
The gpviz package is created to visualize the distribution of the first derivative of the gaussian process predictions.
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.17.3. http://mc-stan.org
}
