% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centralidad.R
\name{mtc_barras}
\alias{mtc_barras}
\title{Grafico de medidas de tendencia central}
\usage{
mtc_barras(bd, x, y, color_base = "#912F40")
}
\arguments{
\item{bd}{data frame de entrada.}

\item{x}{es la variable dependiente, eje y}

\item{y}{es la variable independiente, eje x, a ser la frecuencia}

\item{color_base}{el color del gr\'{a}fico, por defecto es color Gerencia del poder.}
}
\value{
el gr\'{a}fico de medida central como un data frame con la variable dentro del grupo.
}
\description{
Grafico de medidas de tendencia central
}
\section{Quasiquotation}{


\code{graficar_mtc()} es una funci\'{o}n \link[rlang:nse-defuse]{quoting function}. Esto significa
que se utiliza en el contexto del preprocesamiento lo que facilita el trabajo con variables del
data frame graficando sus medidas de tendencia central.
}

\examples{
mtc_barras(mtcars, x = mpg, y = wt)
}
\author{
Gerardo Orantes Jordan, \email{gorantes@gerenciadelpoder.com.mx}
}
