% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frecuencia.R
\name{calcular_frecuencia}
\alias{calcular_frecuencia}
\title{Title}
\usage{
calcular_frecuencia(bd, variables, n_niveles = 9, otro = "Otro", multiples = F)
}
\arguments{
\item{otro}{}
}
\value{

}
\description{
Title
}
