% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frecuencia.R
\name{graficar_frecuencia}
\alias{graficar_frecuencia}
\title{Graficar frecuencias estilo GP}
\usage{
graficar_frecuencia(
  bd,
  variable,
  frecuencia = n,
  n_niveles = 9,
  otro = "Otro",
  grafico = "barras",
  color_base = "#912F40"
)
}
\arguments{
\item{bd}{base de datos}

\item{color_base}{}
}
\value{

}
\description{
Graficar frecuencias estilo GP
}
