% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centralidad.R
\name{graficar_mtc}
\alias{graficar_mtc}
\title{Title}
\usage{
graficar_mtc(
  bd,
  variable,
  grupo,
  medida = "media",
  grafico = "barras",
  color_base = "#912F40"
)
}
\arguments{
\item{color_base}{}
}
\value{

}
\description{
Title
}
